Moved hashable comformance into extension folder.
